Refactor SinglePublisher class out of access token package
